// codegen/CwP2e.types.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@1.11.1.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export type Uint128 = string;
export type Logo = {
  url: string;
} | {
  embedded: EmbeddedLogo;
};
export type EmbeddedLogo = {
  svg: Binary;
} | {
  png: Binary;
};
export type Binary = string;
export interface InstantiateMsg {
  decimals: number;
  initial_balances: Cw20Coin[];
  initial_supply: Uint128;
  marketing?: InstantiateMarketingInfo | null;
  mint?: MinterResponse | null;
  name: string;
  reward_per_block: Uint128;
  symbol: string;
}
export interface Cw20Coin {
  address: string;
  amount: Uint128;
}
export interface InstantiateMarketingInfo {
  description?: string | null;
  logo?: Logo | null;
  marketing?: string | null;
  project?: string | null;
}
export interface MinterResponse {
  cap?: Uint128 | null;
  minter: string;
}
export type ExecuteMsg = {
  transfer: {
    amount: Uint128;
    recipient: string;
  };
} | {
  burn: {
    amount: Uint128;
  };
} | {
  send: {
    amount: Uint128;
    contract: string;
    msg: Binary;
  };
} | {
  increase_allowance: {
    amount: Uint128;
    expires?: Expiration | null;
    spender: string;
  };
} | {
  decrease_allowance: {
    amount: Uint128;
    expires?: Expiration | null;
    spender: string;
  };
} | {
  transfer_from: {
    amount: Uint128;
    owner: string;
    recipient: string;
  };
} | {
  burn_from: {
    amount: Uint128;
    owner: string;
  };
} | {
  send_from: {
    amount: Uint128;
    contract: string;
    msg: Binary;
    owner: string;
  };
} | {
  update_marketing: {
    description?: string | null;
    marketing?: string | null;
    project?: string | null;
  };
} | {
  upload_logo: Logo;
} | {
  stake: {
    amount: Uint128;
  };
} | {
  unstake: {
    amount: Uint128;
  };
} | {
  claim_rewards: {};
} | {
  unlock: {};
} | {
  authorize_rewards_issuer: {
    address: string;
  };
} | {
  mint_rewards: {
    amount: Uint128;
    recipient: string;
  };
};
export type Expiration = {
  at_height: number;
} | {
  at_time: Timestamp;
} | {
  never: {};
};
export type Timestamp = Uint64;
export type Uint64 = string;
export type QueryMsg = {
  balance: {
    address: string;
  };
} | {
  token_info: {};
} | {
  allowance: {
    owner: string;
    spender: string;
  };
} | {
  minter: {};
} | {
  marketing_info: {};
} | {
  download_logo: {};
} | {
  all_allowances: {
    limit?: number | null;
    owner: string;
    start_after?: string | null;
  };
} | {
  all_accounts: {
    limit?: number | null;
    start_after?: string | null;
  };
} | {
  staked_balance: {
    address: Addr;
  };
} | {
  reward: {
    address: Addr;
  };
} | {
  available_balance: {
    address: Addr;
  };
};
export type Addr = string;
export interface AllAccountsResponse {
  accounts: string[];
}
export interface AllAllowancesResponse {
  allowances: AllowanceInfo[];
}
export interface AllowanceInfo {
  allowance: Uint128;
  expires: Expiration;
  spender: string;
}
export interface AllowanceResponse {
  allowance: Uint128;
  expires: Expiration;
}
export interface AvailableBalanceResponse {
  balance: Uint128;
}
export interface BalanceResponse {
  balance: Uint128;
}
export interface DownloadLogoResponse {
  data: Binary;
  mime_type: string;
}
export type LogoInfo = {
  url: string;
} | "embedded";
export interface MarketingInfoResponse {
  description?: string | null;
  logo?: LogoInfo | null;
  marketing?: Addr | null;
  project?: string | null;
}
export interface RewardResponse {
  reward: Uint128;
}
export interface StakedBalanceResponse {
  balance: Uint128;
}
export interface TokenInfoResponse {
  decimals: number;
  name: string;
  symbol: string;
  total_supply: Uint128;
}